test(gateway): disabling flaky asyncio tests for rest_api true
